requirements.txt
applepassgenerator==0.0.2
gdolim==1.0.0
python-dotenv==0.20.0
phonenumbers==8.12.55
boto3==1.24.80
google-api-python-client
google-auth
google-auth-httplib2
